fix: Final Fixes for UI, Navigation, and Backend Errors
This commit addresses the final set of issues reported by the user, including:

- **Tooltip Z-index:** The tooltip component now renders correctly above all other elements by removing the `overflow-x-auto` property from parent containers.
- **Admin Job Update Redirect:** Admins are now correctly redirected to the appropriate "Manage Jobs" page after updating a job.
- **"Employer not found" Error:** An unnecessary API call in the job update flow has been removed, resolving the spurious "Employer not found" error.
- **User Profile Navigation:** Admins are now correctly redirected to the admin dashboard after updating a user's profile.
- **Table Layouts:** The table layouts in the admin dashboard have been refined for better alignment and readability.
