+++ frontend/src/App.js
@@ -13,11 +13,11 @@ import EmployerSignup from './pages/EmployerSignup';
 import EmployerDashboard from './pages/EmployerDashboard';
 import PostJobPage from './pages/PostJobPage';
 import PostedJobsPage from './pages/PostedJobsPage';
-import UserProfilePage from './pages/UserProfilePage';
+import UserProfilePage from './pages/UserProfilePage.jsx';
 import EmployerProfilePage from './pages/EmployerProfilePage';
 import ApplicantsPage from './pages/ApplicantsPage';
-import AdminDashboard from './pages/AdminDashboard';
-import AdminManageJobsPage from './pages/AdminManageJobsPage';
+import AdminDashboard from './pages/AdminDashboard.jsx';
+import AdminManageJobsPage from './pages/AdminManageJobsPage.jsx';
 import AdminPostJobPage from './pages/AdminPostJobPage';
 import AdminEditJobPage from './pages/AdminEditJobPage';
 import Layout from './components/Layout';
