+++ frontend/src/components/common/Tooltip.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Tooltip = ({ text, children }) => {
+  return (
+    <div className="relative flex items-center group">
+      {children}
+      <div className="absolute bottom-full mb-2 w-max px-2 py-1 bg-gray-800 text-white text-xs rounded opacity-0 group-hover:opacity-100 transition-opacity duration-300 pointer-events-none z-50">
+        {text}
+      </div>
+    </div>
+  );
+};
+
+export default Tooltip;
