+++ frontend/src/pages/UserProfilePage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useContext } from 'react';
-import { useParams } from 'react-router-dom';
+import { useParams, useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { getUserProfile, updateUserProfile, getUserById } from '../services/api';
 import { AuthContext } from '../context/AuthContext';
@@ -20,6 +20,7 @@ const UserProfilePage = () => {
   const [loading, setLoading] = useState(true);
   const { updateUser } = useContext(AuthContext);
   const { userId } = useParams();
+  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchProfile = async () => {
@@ -101,12 +102,16 @@ const UserProfilePage = () => {
       // Only update the context if the user is editing their own profile
       if (!userId) {
         // Manually construct the name for the Navbar update
-      const updatedUserForContext = {
-        ...data,
-        name: `${data.firstName} ${data.lastName}`,
-      };
-      updateUser(updatedUserForContext);
-      toast.success('Profile updated successfully!');
+        const updatedUserForContext = {
+          ...data,
+          name: `${data.firstName} ${data.lastName}`,
+        };
+        updateUser(updatedUserForContext);
+        toast.success('Profile updated successfully!');
+      } else {
+        toast.success('User profile updated successfully by admin!');
+        navigate('/admin/dashboard');
+      }
     } catch (error) {
       console.error('Failed to update profile', error);
       toast.error('Failed to update profile.');
