+++ frontend/src/services/api.js
@@ -199,8 +199,11 @@ export const getAdminStats = async (token) => {
 
 export const getAllUsers = async (token, params = {}) => {
   try {
-    const query = new URLSearchParams(params).toString();
-    const response = await fetch(`${API_URL}/admin/users?${query}`, {
+    const query = new URLSearchParams();
+    if (params.search) query.set('search', params.search);
+    if (params.status) query.set('status', params.status);
+    if (params.sortBy) query.set('sortBy', params.sortBy);
+    const response = await fetch(`${API_URL}/admin/users?${query.toString()}`, {
       headers: { Authorization: `Bearer ${token}` },
     });
     if (!response.ok) throw new Error('Failed to fetch users');
@@ -234,9 +237,13 @@ export const getUserById = async (userId, token) => {
   }
 };
 
-export const getAllEmployers = async (token) => {
+export const getAllEmployers = async (token, params = {}) => {
   try {
-    const response = await fetch(`${API_URL}/admin/employers`, {
+    const query = new URLSearchParams();
+    if (params.search) query.set('search', params.search);
+    if (params.status) query.set('status', params.status);
+    if (params.sortBy) query.set('sortBy', params.sortBy);
+    const response = await fetch(`${API_URL}/admin/employers?${query.toString()}`, {
       headers: { Authorization: `Bearer ${token}` },
     });
     if (!response.ok) throw new Error('Failed to fetch employers');
